setup log at children level of click
